+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-* Fixed out of index item creation in pool
+* Fixed out of index item creation in `internal/pool.Pool`
 * Fixed tracing of `(*grpcClientStream).finish` event
 
 ## v3.76.4